Move `WeekdayCondition` to domain
Moves the `WeekdayCondition` enum from the `flow_engine` module to the `domain` module.

This change promotes better separation of concerns by defining `WeekdayCondition`
alongside other domain-specific types, enhancing code organization and reducing
dependencies within the `flow_engine`.
