+++ src/flow_engine/expression.rs
@@ -1,5 +1,5 @@
 use crate::domain::property::{BooleanProperty, NumberProperty, PropertyType};
-use crate::domain::{Number, Weekday};
+use crate::domain::{Number, Weekday, WeekdayCondition};
 use crate::extensions::date_time_ext::ToWeekday;
 use crate::flow_engine::Context;
 use crate::flow_engine::expression::ExpressionError::UnknownProperty;
@@ -55,33 +55,6 @@ pub enum TemporalExpression {
     HasSunSet,   // Now >= sunset
     IsDaytime,   // Now between sunrise and sunset
     IsNighttime, // Now < sunrise or now > sunset
-}
-
-#[derive(Clone, PartialEq, Debug)]
-pub enum WeekdayCondition {
-    Specific(Weekday),
-    Range { start: Weekday, end: Weekday },
-    Set(Vec<Weekday>),
-    Weekdays,
-    Weekend,
-}
-
-impl WeekdayCondition {
-    pub fn included_days(&self) -> Vec<Weekday> {
-        match self {
-            WeekdayCondition::Specific(day) => vec![day.clone()],
-            WeekdayCondition::Range { start, end } => {
-                let all = Weekday::all();
-                let start_index = start.as_index();
-                let end_index = end.as_index();
-
-                all[start_index..=end_index].to_vec()
-            }
-            WeekdayCondition::Set(days) => days.clone(),
-            WeekdayCondition::Weekdays => vec![Monday, Tuesday, Wednesday, Thursday, Friday],
-            WeekdayCondition::Weekend => vec![Saturday, Sunday],
-        }
-    }
 }
 
 #[derive(PartialEq, Debug, Clone)]
