+++ src/flow_loader/weekday_condition_deserializer.rs
@@ -1,5 +1,4 @@
-use crate::domain::Weekday;
-use crate::flow_engine::WeekdayCondition;
+use crate::domain::{Weekday, WeekdayCondition};
 use serde::de::{SeqAccess, Unexpected, Visitor};
 use serde::{Deserialize, Deserializer, de};
 use std::fmt::Formatter;
